requirements.txt
streamlit==1.45.1
openai==1.78.1
python-dotenv==1.1.0
chardet==5.2.0
st_supabase_connection==1.2.0
cryptography==42.0.5
sentence-transformers==4.1.0
chromadb==1.0.12
numpy==1.25.2
torch==2.1.2
resend==2.10.0
backoff==2.2.1
streamlit-cookies-controller==0.0.4